Improve documentation of the `position` function.
Improve the style of the summary of the function and add note about
surface position message and reference position. Describe reference
position arguments.

Improve formatting of the error messages in the exceptions raised by the
function (i.e. it seems most of the error messages start with a capital
letter).